+++ src/main.cpp
@@ -4,45 +4,110 @@ unsigned int K;
 
 BEGIN_ALLOC_FUNCTION(FirstFit)
 {
-  int numberOfSlots = REQ_SLOTS(0);
   int currentNumberSlots;
   int currentSlotIndex;
+  int routeLength;
+  int requiredSlots;
   std::vector<bool> totalSlots;
-  for (int i = 0; i < NUMBER_OF_ROUTES;
-       i++)
-  {
-    if (i >= K) continue;
-    totalSlots = std::vector<bool>(LINK_IN_ROUTE(0, 0)->getSlots(),
-                                   false);
-    for (int j = 0; j < NUMBER_OF_LINKS(i);
-         j++)
-    {
-      for (int k = 0; k < LINK_IN_ROUTE(i, j)->getSlots();
-           k++)
-      {
-        totalSlots[k] = totalSlots[k] | LINK_IN_ROUTE(i, j)->getSlot(k);
+ 
+  for (int r = 0; r < NUMBER_OF_ROUTES; r++) {
+ 
+    routeLength = 0;
+    totalSlots = std::vector<bool>(LINK_IN_ROUTE(r, 0)->getSlots(), false);
+    for (int l = 0; l < NUMBER_OF_LINKS(r); l++) {
+      routeLength += LINK_IN_ROUTE(r, l)->getLength();
+      for (int s = 0; s < LINK_IN_ROUTE(r, l)->getSlots(); s++) {
+        totalSlots[s] = totalSlots[s] | LINK_IN_ROUTE(r, l)->getSlot(s);
       }
     }
-    currentNumberSlots = 0;
-    currentSlotIndex = 0;
-    for (int j = 0; j < totalSlots.size();
-         j++)
-    {
-      if (totalSlots[j] == false)
-      {
-        currentNumberSlots++;
+ 
+    for (int m = 0; m < NUMBER_OF_MODULATIONS; m++){
+ 
+      if (routeLength > REQ_REACH(m)) continue;
+ 
+      requiredSlots = REQ_SLOTS(m);
+      currentNumberSlots = 0;
+      currentSlotIndex = 0;
+ 
+      for (int s = 0; s < totalSlots.size(); s++) {
+        if (totalSlots[s] == false) {
+          currentNumberSlots++;
+        }
+        else {
+          currentNumberSlots = 0;
+          currentSlotIndex = s + 1;
+        }
+        if (currentNumberSlots == requiredSlots) {
+          for (int l = 0; l < NUMBER_OF_LINKS(r); l++) {
+            ALLOC_SLOTS(LINK_IN_ROUTE_ID(r, l), currentSlotIndex, requiredSlots);
+          }
+          return ALLOCATED;
+        }
       }
-      else
-      {
-        currentNumberSlots = 0;
-        currentSlotIndex = j + 1;
+    }
+  }
+  return NOT_ALLOCATED;
+}
+END_ALLOC_FUNCTION
+
+BEGIN_ALLOC_FUNCTION(BestFit) {
+  int currentNumberSlots;
+  int currentSlotIndex;
+  int routeLength;
+  int requiredSlots;
+  int bestTotal;
+  int bestSlotIndex;
+  std::vector<bool> totalSlots;
+ 
+  for (int r = 0; r < NUMBER_OF_ROUTES; r++) {
+ 
+    routeLength = 0;
+    totalSlots = std::vector<bool>(LINK_IN_ROUTE(r, 0)->getSlots(), false);
+    for (int l = 0; l < NUMBER_OF_LINKS(r); l++) {
+      routeLength += LINK_IN_ROUTE(r, l)->getLength();
+      for (int s = 0; s < LINK_IN_ROUTE(r, l)->getSlots(); s++) {
+        totalSlots[s] = totalSlots[s] | LINK_IN_ROUTE(r, l)->getSlot(s);
       }
-      if (currentNumberSlots == numberOfSlots)
-      {
-        for (int j = 0; j < NUMBER_OF_LINKS(i);
-             j++)
-        {
-          ALLOC_SLOTS(LINK_IN_ROUTE_ID(i, j), currentSlotIndex, numberOfSlots)
+    }
+
+    for (int m = 0; m < NUMBER_OF_MODULATIONS; m++){
+
+      if (routeLength > REQ_REACH(m)) continue;
+
+      requiredSlots = REQ_SLOTS(m);
+      bestTotal = std::numeric_limits<int>::max();
+      bestSlotIndex = -1;
+      currentNumberSlots = 0;
+      currentSlotIndex = 0;
+
+      for (int s = 0; s < totalSlots.size(); s++) {
+        if (!totalSlots[s]) {
+          currentNumberSlots++;
+        }
+        else {
+          if (currentNumberSlots == requiredSlots) {
+            for (int l = 0; l < NUMBER_OF_LINKS(r); l++) {
+              ALLOC_SLOTS(LINK_IN_ROUTE_ID(r, l), currentSlotIndex, requiredSlots);
+            }
+            return ALLOCATED;
+          }
+          else if (currentNumberSlots >= requiredSlots && currentNumberSlots < bestTotal) {
+            bestTotal = currentNumberSlots;
+            bestSlotIndex = currentSlotIndex;
+          }
+          currentNumberSlots = 0;
+          currentSlotIndex = s + 1;
+        }
+      }
+
+      if (currentNumberSlots >= requiredSlots && currentNumberSlots < bestTotal) {
+        bestTotal = currentNumberSlots;
+        bestSlotIndex = currentSlotIndex;
+      }
+
+      if (bestSlotIndex != -1) {
+        for (int l = 0; l < NUMBER_OF_LINKS(r); l++) {
+          ALLOC_SLOTS(LINK_IN_ROUTE_ID(r, l), bestSlotIndex, requiredSlots);
         }
         return ALLOCATED;
       }
@@ -52,89 +117,24 @@ BEGIN_ALLOC_FUNCTION(FirstFit)
 }
 END_ALLOC_FUNCTION
 
-BEGIN_ALLOC_FUNCTION(ExactFit) {
-  int numberOfSlots = REQ_SLOTS(0);
-  int currentNumberSlots;
-  int currentSlotIndex;
-  int firstIndex;
-  std::vector<bool> totalSlots;
-  for (int i = 0; i < NUMBER_OF_ROUTES; i++) {
-    if (i >= K) continue;
-    totalSlots = std::vector<bool>(LINK_IN_ROUTE(0, 0)->getSlots(), false);
-    firstIndex = -1;
-    for (int j = 0; j < NUMBER_OF_LINKS(i); j++) {
-      for (int k = 0; k < LINK_IN_ROUTE(i, j)->getSlots(); k++) {
-        totalSlots[k] = totalSlots[k] | LINK_IN_ROUTE(i, j)->getSlot(k);
-      }
-    }
-    currentNumberSlots = 0;
-    currentSlotIndex = 0;
-    for (int j = 0; j < totalSlots.size(); j++) {
-      if (totalSlots[j] == false) {
-        currentNumberSlots++;
-      } else {
-        if (currentNumberSlots == numberOfSlots) {
-          for (int j = 0; j < NUMBER_OF_LINKS(i); j++) {
-            ALLOC_SLOTS(LINK_IN_ROUTE_ID(i, j), currentSlotIndex, numberOfSlots)
-          }
-          return ALLOCATED;
-        }
-        currentNumberSlots = 0;
-        currentSlotIndex = j + 1;
-      }
-      if (firstIndex == -1 && currentNumberSlots > numberOfSlots) {
-        firstIndex = currentSlotIndex;
-      }
-    }
-    if (firstIndex != -1) {
-      for (int j = 0; j < NUMBER_OF_LINKS(i); j++) {
-        ALLOC_SLOTS(LINK_IN_ROUTE_ID(i, j), firstIndex, numberOfSlots);
-      }
-      return ALLOCATED;
-    }
-  }
-  return NOT_ALLOCATED;
-}
-END_ALLOC_FUNCTION
-
 int main(int argc, char *argv[])
 {
-
   if (argc < 10) {
     std::cerr << "Uso: " << argv[0] << " <AlgorithmName> <networkType> <goalConnections> <confidence> <lambda> <mu> <networkName> <bitrate> <K>" << std::endl;
     return 1;
   }
 
   int networkType = std::stoi(argv[2]);
-  if (networkType != 1) {
-    std::cerr << "At the moment only networkType 1 is supported" << std::endl;
-    return 1;
-  }
-
   int goalConnections = std::stoi(argv[3]);
-  if (goalConnections > 10000000) {
-    std::cerr << "goalConnections must be less than 10,000,000" << std::endl;
-    return 1;
-  } else if (goalConnections < 1) {
-    std::cerr << "goalConnections must be greater than 0" << std::endl;
-    return 1;
-  }
-
   float confidence = std::stof(argv[4]);
   float lambda = std::stof(argv[5]);
-
   float mu = std::stof(argv[6]);
   K = std::stoi(argv[9]);
-  if (K > 6) {
-    std::cerr << "Max K is 6" << std::endl;
-    return 1;
-  } else if (K < 1) {
-    std::cerr << "Min K is 1" << std::endl;
-    return 1;
-  }
-
   std::string networkName = argv[7];
   std::string bitrate = argv[8];
+  
+  // We're no longer doing validation here as it's handled by the API layer
+  
   Simulator sim(
       std::string("./networks/" + networkName + ".json"),
       std::string("./networks/" + networkName + "_routes.json"),
@@ -148,23 +148,28 @@ int main(int argc, char *argv[])
     USE_ALLOC_FUNCTION(FirstFit, sim);
     break;
 
-  case 'E':
-    USE_ALLOC_FUNCTION(ExactFit, sim);
+  case 'B':
+    USE_ALLOC_FUNCTION(BestFit, sim);
     break;
 
   default:
+    // Still keep this validation for safety
     std::cerr << "Invalid algorithm" << std::endl;
     return 1;
     break;
   }
 
-  USE_ALLOC_FUNCTION(FirstFit, sim);
   sim.setGoalConnections(goalConnections);
   sim.setConfidence(confidence);
   sim.setLambda(lambda);
   sim.setMu(mu);
   sim.init();
   sim.run();
+
+  // Print the results with flush
+  // Set the precision to 6 decimal places
+  std::cout.precision(4);
+  std::cout << "final_blocking:   " << sim.getBlockingProbability() << "\n" << std::flush;
 
   return 0;
 }